ui: create unique component only for tag editing
